refactor: convert current speed to float

+++ driving/api/src/main/kotlin/com/krilatokolo/wingeddriver/driving/ActiveLocoState.kt
@@ -2,7 +2,7 @@ package com.krilatokolo.wingeddriver.driving
 
 data class ActiveLocoState(
    val id: Int,
-   val speed: Int,
+   val speed: Float,
    val forward: Boolean,
    val maxSpeed: Int,
    val activeFunctions: List<Int>,
